// Generador/front/src/dopplerApi/types/Report.tsx
// Generated by https://quicktype.io

import { Measure } from "./Measures";
import { Neighborhood } from "./Neighborhoods";
import { Radar } from "./Radars";

export type Report = {
  actNumber: string;
  date: string | Date;
  speed: number;
  plate: string;
  location: string;
  neighborhood: Neighborhood;
  radar: Radar;
  measure: string | Measure;
  src: Src;
  homeowner?: string;
  functionalUnit?: string;
  _id: string;
  __v: number;
};

export type ReportWithMeasure = Report & {
  measure: Measure;
};

export interface Src {
  car: string;
  domain?: string;
}